+++ FootballBet.Infrastructure/Interfaces/IBalanceService.cs
@@ -2,6 +2,6 @@ namespace FootballBet.Infrastructure.Interfaces;
 
 public interface IBalanceService
 {
-    public Task<double> GetBalance(string userId, CancellationToken cancellationToken = default);
-    public Task<double> UpdateBalance(string userId, double amount, CancellationToken cancellationToken = default);
+    public Task<decimal> GetBalance(string userId, CancellationToken cancellationToken = default);
+    public Task<decimal> UpdateBalance(string userId, decimal amount, CancellationToken cancellationToken = default);
 }
